requirements.txt
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0